feat: move MITM core logic into proxy-sidecar package
