+++ proxy-sidecar/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@vellumai/proxy-sidecar",
+  "version": "0.1.0",
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "scripts": {
+    "typecheck": "bunx tsc --noEmit",
+    "test": "bun test"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@types/bun": "^1.2.4",
+    "@types/node": "^25.2.2",
+    "typescript": "^5.7.3"
+  }
+}
